+++ marketing/src/containers/compliance/video/component.tsx
@@ -1,39 +1,10 @@
-import { useEffect, useRef } from 'react';
-import { useInView } from 'react-intersection-observer';
+import { FC } from 'react';
+import Video from 'containers/compliance/video';
 
-import { Media } from 'components/media-query';
-import Image from 'next/image';
+const Webinar: FC = () => (
+  <div className="relative w-full min-h-[495px] space-y-12 py-12 px-14 mx-auto ">
+    <Video />
+  </div>
+);
 
-const Video: React.FC = () => {
-  const videoRef = useRef<HTMLVideoElement>();
-  const { ref, inView } = useInView();
-
-  useEffect(() => {
-    if (videoRef.current && inView) {
-      videoRef.current.play();
-    }
-
-    if (videoRef.current && !inView) {
-      videoRef.current.pause();
-    }
-  }, [inView]);
-
-  return (
-    <section ref={ref} className="relative max-h-screen bg-white">
-      <Media lessThan="sm">
-        <Image
-          layout="responsive"
-          src="/images/about/video/tractorimage.jpg"
-          alt="Video tractor image"
-          width={375}
-          height={211}
-        />
-      </Media>
-      <Media greaterThanOrEqual="md">
-        <video ref={videoRef} src="/videos/truck.mp4" className="aspect-video" loop muted />
-      </Media>
-    </section>
-  );
-};
-
-export default Video;
+export default Webinar;
