--- marketing/src/containers/compliance/video/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
