--- marketing/src/containers/compliance/webinar/component-variant.tsx
@@ -1,10 +1,0 @@
-import { FC } from 'react';
-import Video from 'containers/compliance/video';
-
-const Webinar: FC = () => (
-  <div className="relative w-full min-h-[495px] space-y-12 py-12 px-14 mx-auto ">
-    <Video />
-  </div>
-);
-
-export default Webinar;
